feat(web): modularize transaction detail sections
